Updates for tagging and filenaming

+++ bin/SpotifyAPI.py
@@ -1,6 +1,7 @@
-import json, os, math
-from Naked.toolshed.shell import muterun 
-from mutagen.mp4 import MP4, MP4Cover
+import requests, sys, getopt, json, math, os
+from Naked.toolshed.shell import muterun
+from mutagen.mp4 import MP4, MP4Cover, MP4StreamInfoError
+from time import sleep
 
 class SpotifyTrack:
     name = ""
@@ -12,17 +13,17 @@ class SpotifyTrack:
     artwork_url = ""
     uri = ""
 
-    def getDurationSeconds(self):
-    	return int(math.ceil(self.duration / 1000.0))
+    def getDurationSeconds(self, n=0):
+    	return int(math.ceil(self.duration / 1000.0)) + n
 
     # process the JSON for this Spotify Track
     def processJson(self, j):
         self.uri = j["track"]["uri"]
-    	self.name = j["track"]["name"]
+    	self.name = j["track"]["name"].replace('/', ':')
     	self.duration = j["track"]["duration_ms"]
     	self.track_num = j["track"]["track_number"]
     	self.artwork_url = j["track"]["album"]["images"][0]["url"]
-    	self.album = j["track"]["album"]["name"] 
+    	self.album = j["track"]["album"]["name"]
     	self.artist = j["track"]["artists"][0]["name"]
     	self.disc_num = j["track"]["disc_number"]
 
@@ -33,20 +34,30 @@ class SpotifyTrack:
     		return False
     	return True
 
-    def tag(self, filename):
-    	f = MP4(filename)
-    	f.tags['\xa9nam'] = self.name
-    	f.tags['\xa9alb'] = self.album
-    	f.tags['\xa9ART'] = self.artist
-    	f.tags['trkn'] = [(self.track_num,self.track_num)]
-    	f.tags['disk'] = [(self.disc_num,self.disc_num)]
-    	if self.loadArtwork(self.artwork_url):
-    		with open("/tmp/cover.jpeg", "rb") as art:
-    			f.tags["covr"] = [
-        			MP4Cover(art.read(), imageformat=MP4Cover.FORMAT_JPEG)
-    			]
-    		os.remove('/tmp/cover.jpeg')
-    	f.save()
+    def sanitize(self, filename):
+        return filename.replace('/', ':')
+
+    def tag(self, filename, withRetry=True):
+        if os.path.isfile(filename):
+            # tag the track, need to loop because the export process in non-deterministic
+            while withRetry:
+                try:
+                    f = MP4(filename)
+                    f.tags['\xa9nam'] = self.name
+                    f.tags['\xa9alb'] = self.album
+                    f.tags['\xa9ART'] = self.artist
+                    f.tags['trkn'] = [(self.track_num,self.track_num)]
+                    f.tags['disk'] = [(self.disc_num,self.disc_num)]
+                    if self.loadArtwork(self.artwork_url):
+                        with open("/tmp/cover.jpeg", "rb") as art:
+                            f.tags["covr"] = [
+                                MP4Cover(art.read(), imageformat=MP4Cover.FORMAT_JPEG)
+                            ]
+                        os.remove('/tmp/cover.jpeg')
+                    f.save()
+                    break
+                except MP4StreamInfoError:
+                    sleep(1)
 
     def __init__(self, j):
     	self.processJson(j)
